requirements.txt
psycopg2-binary==2.9.5
python-decouple==3.6
SQLAlchemy==1.4.44
pytest==7.2.0
